Create a method called update_position
Created a method called update_position in Board
class. This method calls two other methods i.e.
new_position? method and previous_position? method
These two methods will ask user which chess piece
to move and to where.

This was done by getting the chess notation from
the user such as e5 or a6 and splitting them up
into an array and reverse the order. After
some math because of 0-base indexing. We finally
get the row and column needed to update the
position of the chess piece. The previous position
of the chess piece becomes eempty therefore a '.'
is placed.