Add: legal_moves rook class, check empty moves.
Added legal_moves method to rook class which
first checks if there is an empty space in front
of the rook and then counts down and space until
a chess piece is found. This stops the legal_moves
method from getting coordinates that are behind a
chess piece.

Also added a guard clause for when, if the chess
piece chosen by user has no moves. He has to
choose another piece.